Add files via upload
This is a Jupyter notebook where I implemented a medical image segmentation project using the U-Net architecture. I worked on this project to familiarize myself with the fundamentals of medical image segmentation and the U-Net model, which I applied later during my internship at PINUM Cancer Hospital for tumor segmentation tasks. The goal of this project was to explore the capabilities of U-Net in segmenting medical images and prepare myself for more advanced work in tumor segmentation. The dataset used consists of ultrasound images, and the project demonstrates techniques in image preprocessing, model training, and evaluation.